File moved to Utilities -> Util .hpp/cpp
