+++ src/collision/pchCollision.hpp
@@ -9,7 +9,7 @@ This file is released under the MIT license. See README-MIT for more information
 #include <sstream>
 #include <iomanip>
 #include <Common.hpp>
-#include <Util.hpp>
+#include <Utilities/Util.hpp>
 #include <Logging/Logger.hpp>
 #include <Models/ModelInstance.h>
 #include <Models/GameObjectModel.h>
