+++ src/scripts/SpellHandlers/CMakeLists.txt
@@ -14,7 +14,7 @@ set(sources
     # required for linker
     ../../world/Server/ServerState.cpp
     ../../world/Spell/SpellCastTargets.cpp
-    ../../shared/Util.cpp
+    ../../shared/Utilities/Util.cpp
 )
 
 set(headers
