+++ src/shared/Config/Config.cpp
@@ -6,7 +6,7 @@ This file is released under the MIT license. See README-MIT for more information
 #include "Common.hpp"
 #include "Config.h"
 #include "Logging/Logger.hpp"
-#include "Util.hpp"
+#include "Utilities/Util.hpp"
 #include <stdexcept>
 
 bool ConfigFile::openAndLoadConfigFile(const std::string& configFileName)
