+++ src/shared/Database/Database.cpp
@@ -26,7 +26,7 @@
  //////////////////////////////////////////////
 
 #include "DatabaseCommon.hpp"
-#include "Util.hpp"
+#include "Utilities/Util.hpp"
 #include <string>
 #include <vector>
