+++ src/shared/Threading/LegacyThreadPool.cpp
@@ -19,7 +19,7 @@
 
 #include "LegacyThreadPool.h"
 #include <Logging/Logger.hpp>
-#include "Util.hpp"
+#include "Utilities/Util.hpp"
 
 #ifdef WIN32
