+++ src/shared/Utilities/CMakeLists.txt
@@ -5,6 +5,8 @@ set(PATH_PREFIX Utilities)
 set(SRC_UTILITIES_FILES
     ${PATH_PREFIX}/Strings.cpp
     ${PATH_PREFIX}/Strings.hpp
+    ${PATH_PREFIX}/Util.cpp
+    ${PATH_PREFIX}/Util.hpp
 )
 
 source_group(Utilities FILES ${SRC_UTILITIES_FILES})
