+++ src/tools/mmaps_generator/PathGenerator.cpp
@@ -21,7 +21,7 @@
 
 #include "PathCommon.h"
 #include "MapBuilder.h"
-#include "Util.hpp"
+#include "Utilities/Util.hpp"
 
 namespace fs = std::filesystem;
