+++ src/world/Movement/MovementGenerators/ChaseMovementGenerator.cpp
@@ -12,7 +12,7 @@ This file is released under the MIT license. See README-MIT for more information
 #include "Movement/Spline/MoveSplineInit.h"
 #include "Movement/PathGenerator.h"
 #include "Objects/Units/Unit.hpp"
-#include "Util.hpp"
+#include "Utilities/Util.hpp"
 
 static bool hasLostTarget(Unit* owner, Unit* target)
 {
