+++ src/world/Movement/MovementGenerators/FormationMovementGenerator.h
@@ -8,7 +8,7 @@ This file is released under the MIT license. See README-MIT for more information
 #include "Movement/AbstractFollower.h"
 #include "Movement/MovementGenerator.h"
 #include "LocationVector.h"
-#include "Util.hpp"
+#include "Utilities/Util.hpp"
 
 class Creature;
