+++ src/world/Movement/MovementGenerators/WaypointMovementGenerator.h
@@ -7,7 +7,7 @@ This file is released under the MIT license. See README-MIT for more information
 
 #include "Movement/MovementGenerator.h"
 #include "PathMovementBase.h"
-#include "Util.hpp"
+#include "Utilities/Util.hpp"
 
 class Creature;
 class Unit;
