+++ src/world/Movement/WaypointManager.cpp
@@ -4,7 +4,7 @@ This file is released under the MIT license. See README-MIT for more information
 */
 
 #include "WaypointManager.h"
-#include "Util.hpp"
+#include "Utilities/Util.hpp"
 #include "Logging/Logger.hpp"
 #include "Server/MainServerDefines.h"
 #include "Storage/MySQLDataStore.hpp"
