+++ src/world/Server/EventableObject.h
@@ -22,7 +22,7 @@
 #define EVENTABLEOBJECT_H
 
 #include "EventMgr.h"
-#include <Util.hpp>
+#include <Utilities/Util.hpp>
 #include <list>
 #include <set>
