+++ src/world/Spell/Spell.cpp
@@ -48,7 +48,7 @@ This file is released under the MIT license. See README-MIT for more information
 #include "Objects/Units/Creatures/Pet.h"
 #include "Objects/Units/Players/PlayerClasses.hpp"
 #include "Objects/Units/UnitDefines.hpp"
-#include "Util.hpp"
+#include "Utilities/Util.hpp"
 
 using namespace AscEmu::Packets;
